feat: restructure for multi-platform portability + PS2 backend
- Engine is now a pure abstract interface (zero SDL2 in header)
- All SDL2 code moved to src/platform/sdl2/
- New PS2 backend in src/platform/ps2/ (SDL 1.2 + libpad + PS2SDK)
- C++ standard downgraded from C++23 to C++17
- double replaced with float everywhere (PS2 R5900 has no double FPU)
- dynamic_cast eliminated via state_type() virtual method
- Added StateType enum for RTTI-free state identification
- SoundManager volume calls abstracted through Engine interface
- CI workflow: added PS2 build job (ps2dev/ps2dev container)
- Web/Deploy workflows updated to C++17 with new source paths
- Cross file for PS2: cross/ps2.ini
- Build script: build-ps2.sh

+++ src/core/StateMachine.cpp
@@ -12,7 +12,7 @@ auto StateMachine::switch_state_raw(std::unique_ptr<GameState> state) -> void {
     m_current = std::move(state);
 }
 
-auto StateMachine::update(double dt) -> void {
+auto StateMachine::update(float dt) -> void {
     if (m_current) {
         m_current->update(dt);
     }
