+++ src/game1/Constants.hpp
@@ -9,9 +9,9 @@ constexpr int SCREEN_HEIGHT = 720;
 constexpr int FPS = 60;
 constexpr const char* ASSETS_DIR = "assets";
 
-constexpr double HOUR_DURATION = 60.0;
-constexpr double MAX_POWER = 100.0;
-constexpr double BASE_POWER_DRAIN = 0.12;
+constexpr float HOUR_DURATION = 60.0f;
+constexpr float MAX_POWER = 100.0f;
+constexpr float BASE_POWER_DRAIN = 0.12f;
 constexpr int OFFICE_WIDTH = 1280;
 
 inline std::string asset_path(const std::string& sub, const std::string& ext) {
@@ -26,8 +26,8 @@ inline std::string asset_full(const std::string& rel) {
     return std::string(ASSETS_DIR) + "/" + rel;
 }
 
-constexpr double PAN_SPEED = 8.0;
+constexpr float PAN_SPEED = 8.0f;
 constexpr int PAN_MARGIN = 200;
-constexpr double MOVE_INTERVAL = 5.0;
+constexpr float MOVE_INTERVAL = 5.0f;
 
 }  // namespace fnwf::GameSettings
